Move definition of OldGlobalConfig and related functions to globalconfigold.go to keep global.go clean. Avoids confusion when adding new config items.
